# src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: Cooperative Voting API
  description: API de votação cooperativa. Gerencia pautas submetidas à assembleia.
  version: "1.0.0"

servers:
  - url: /

tags:
  - name: Pauta
    description: Cadastro e consulta de pautas de votação
  - name: Sessão de Votação
    description: Abertura de sessões de votação em pautas

paths:
  /pautas:
    post:
      tags: [Pauta]
      summary: Cadastra uma nova pauta
      operationId: criarPauta
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PautaRequest"
      responses:
        "201":
          description: Pauta criada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PautaResponse"
        "400":
          description: Dados inválidos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErroValidacao"
    get:
      tags: [Pauta]
      summary: Lista pautas com paginação e filtros por data de criação
      operationId: listarPautas
      parameters:
        - name: page
          in: query
          description: Número da página (0-indexed)
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          description: Tamanho da página
          schema:
            type: integer
            default: 20
        - name: sort
          in: query
          description: "Campo e direção de ordenação, ex.: dataCriacao,desc"
          schema:
            type: string
        - name: dataCriacaoMaior
          in: query
          description: Filtra pautas criadas após esta data
          schema:
            type: string
            format: date-time
        - name: dataCriacaoMenor
          in: query
          description: Filtra pautas criadas antes desta data
          schema:
            type: string
            format: date-time
      responses:
        "200":
          description: Página de pautas
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PautaPage"

  /pautas/{id}:
    get:
      tags: [Pauta]
      summary: Busca uma pauta pelo id
      operationId: buscarPautaPorId
      parameters:
        - $ref: "#/components/parameters/IdPauta"
      responses:
        "200":
          description: Pauta encontrada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PautaResponse"
        "404":
          description: Pauta não encontrada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Erro"
    put:
      tags: [Pauta]
      summary: Atualiza uma pauta existente
      operationId: atualizarPauta
      parameters:
        - $ref: "#/components/parameters/IdPauta"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PautaRequest"
      responses:
        "200":
          description: Pauta atualizada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PautaResponse"
        "400":
          description: Dados inválidos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErroValidacao"
        "404":
          description: Pauta não encontrada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Erro"
    delete:
      tags: [Pauta]
      summary: Exclui (logicamente) uma pauta existente
      operationId: excluirPauta
      parameters:
        - $ref: "#/components/parameters/IdPauta"
      responses:
        "204":
          description: Pauta excluída
        "404":
          description: Pauta não encontrada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Erro"

  /pautas/{id}/sessoes:
    post:
      tags: [Sessão de Votação]
      summary: Abre uma sessão de votação para a pauta
      description: >-
        Corpo é opcional. Sem `duracaoSegundos` (ou sem corpo), a sessão fica aberta
        por 1 minuto. Só é permitida uma sessão por pauta.
      operationId: abrirSessaoVotacao
      parameters:
        - $ref: "#/components/parameters/IdPauta"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SessaoVotacaoRequest"
      responses:
        "201":
          description: Sessão aberta
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessaoVotacaoResponse"
        "400":
          description: Dados inválidos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErroValidacao"
        "404":
          description: Pauta não encontrada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Erro"
        "409":
          description: Sessão de votação já aberta para a pauta
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Erro"

components:
  parameters:
    IdPauta:
      name: id
      in: path
      required: true
      description: Id (UUID) da pauta
      schema:
        type: string
        format: uuid
      example: 3fa85f64-5717-4562-b3fc-2c963f66afa6

  schemas:
    PautaRequest:
      type: object
      required: [titulo]
      properties:
        titulo:
          type: string
          maxLength: 120
          example: Alterar estatuto social
        descricao:
          type: string
          maxLength: 500
          example: Proposta de alteração do estatuto social da cooperativa

    PautaResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
        titulo:
          type: string
          example: Alterar estatuto social
        descricao:
          type: string
          example: Proposta de alteração do estatuto social da cooperativa
        dataCriacao:
          type: string
          format: date-time
        dataAtualizacao:
          type: string
          format: date-time
          nullable: true

    PautaPage:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/PautaResponse"
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
        number:
          type: integer
          description: Página atual (0-indexed)
        size:
          type: integer
        first:
          type: boolean
        last:
          type: boolean

    Erro:
      type: object
      properties:
        status:
          type: integer
        error:
          type: string
        codigo:
          type: string
          description: Código do erro de negócio (catálogo interno). Ausente em erros de framework/validação.
          nullable: true
          example: PAUTA_NAO_ENCONTRADA
        message:
          type: string
        path:
          type: string

    SessaoVotacaoRequest:
      type: object
      properties:
        duracaoSegundos:
          type: integer
          format: int32
          minimum: 1
          description: Duração da sessão em segundos. Default de 60s (1 minuto) se omitido.
          example: 60

    SessaoVotacaoResponse:
      type: object
      properties:
        pautaId:
          type: string
          format: uuid
          example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
        dataAbertura:
          type: string
          format: date-time
        dataFechamento:
          type: string
          format: date-time
        status:
          type: string
          enum: [ABERTA, ENCERRADA]

    ErroValidacao:
      allOf:
        - $ref: "#/components/schemas/Erro"
        - type: object
          properties:
            errors:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string